requirements.txt
asgiref==3.9.1
Django==5.2.5
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
python-decouple==3.8
redis==6.4.0
sqlparse==0.5.3
tzdata==2025.2
